refactor(press): rename createPressEvent to create-press-event
